neurips_supp/codes/internvideo2_vit/requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
argon2-cffi-bindings==21.2.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.3.0
black==24.4.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.0.0
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.10
datasets==3.5.0
decorator==5.2.1
decord==0.6.0
diffusers==0.26.3
dill==0.3.8
easydict==1.13
einops==0.8.1
exceptiongroup==1.2.1
filelock==3.18.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20221221
grpcio==1.64.1
h11==0.14.0
huggingface-hub==0.30.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==7.0.1
iopath==0.1.9
ipywidgets==8.1.2
isoduration==20.11.0
Jinja2==3.1.6
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.1
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-kernel-gateway==3.0.1
jupyter-lsp==2.2.5
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.8
loralib==0.1.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
moviepy==2.1.2
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
peft==0.15.1
pillow==10.4.0
pip==25.0.1
portalocker==2.8.2
proglog==0.1.11
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.18.1
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==75.8.2
simplejson==3.19.2
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.0.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
trl==0.16.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
warmup-scheduler==0.3
webcolors==1.13
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.45.1
widgetsnbextension==4.0.10
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
youtube-dl==2021.12.17
zipp==3.17.0